// src/enums/media-fields.ts
// -----------------------------------------------------------------------------------------
// #region Enums
// -----------------------------------------------------------------------------------------

enum MediaFields {
    DurationMs = "duration_ms",
    Height = "height",
    MediaKey = "media_key",
    PreviewImageUrl = "preview_image_url",
    Type = "type",
    Url = "url",
    Width = "width",
    PublicMetrics = "public_metrics",
    NonPublicMetrics = "non_public_metrics",
    OrganicMetrics = "organic_metrics",
    PromotedMetrics = "promoted_metrics",
}

// #endregion Enums

// -----------------------------------------------------------------------------------------
// #region Exports
// -----------------------------------------------------------------------------------------

export { MediaFields };

// #endregion Exports